requirements.txt
aiogram==3.13.1
openai==1.53.0
aiohttp==3.10.10
pymongo==4.10.1
python-ffmpeg==2.0.12
yandex-speechkit==1.5.0
grpcio-tools==1.48.2
tiktoken==0.8.0
python-docx==1.1.2
moviepy==1.0.3